rft: 还原菜单相关内容到SettingsViewModel
